+++ language/da/googleanalytics_acp.php
@@ -2,11 +2,12 @@
 /**
 *
 * Google Analytics extension for the phpBB Forum Software package.
-* Translate by: rexkyo (http://phpbb-tw.net/)
 *
 * @copyright (c) 2014 phpBB Limited <https://www.phpbb.com>
 * @license GNU General Public License, version 2 (GPL-2.0)
 *
+* Google Analytics extension [Danish]
+* Translator: Klaus Rasmussen
 */
 
 /**
@@ -40,7 +41,7 @@ if (empty($lang) || !is_array($lang))
 
 $lang = array_merge($lang, array(
 	'ACP_GOOGLEANALYTICS'				=> 'Google Analytics',
-	'ACP_GOOGLEANALYTICS_ID'			=> 'Google Analytics（分析） ID',
-	'ACP_GOOGLEANALYTICS_ID_EXPLAIN'	=> '輸入你的 Google Analytics（分析） ID code, 例如: <samp>UA-0000000-00</samp>.<br /><br />Google Analytics（分析）可以跨多個設備和會話跟踪您的註冊用戶，以便在報告中提供更準確的用戶數。 要啟用此增強功能，您必須在Google Analytics（分析）帳戶中配置用戶ID跟踪。<a href="https://support.google.com/analytics/answer/3123666"> 單擊獲取更多信息。</a>',
-	'ACP_GOOGLEANALYTICS_ID_INVALID'	=> '“%s” 不是有效的 Google Analytics（分析） ID code.<br />它的格式應該會是像這樣 “UA-0000000-00”.',
+	'ACP_GOOGLEANALYTICS_ID'			=> 'Google Analytics ID',
+	'ACP_GOOGLEANALYTICS_ID_EXPLAIN'	=> 'Indtast din Google Analytics ID kode, f.eks.: <samp>UA-0000000-00</samp>.<br /><br />Google Analytics kan spore dine registrerede brugere på tværs af flere enheder og sessioner, for et mere præcist bruger tal i dine rapportere. For at aktivere denne forbedrede funktionalitet skal User ID sporing være konfigureret i din Google Analytics konto. <a href="https://support.google.com/analytics/answer/3123666">Klik for yderligere information</a>.',
+	'ACP_GOOGLEANALYTICS_ID_INVALID'	=> '“%s” er ikke en gyldig Google Analytics ID kode.<br />Det skal være i formen: “UA-0000000-00”.',
 ));