+++ language/ro/googleanalytics_acp.php
@@ -39,7 +39,8 @@ if (empty($lang) || !is_array($lang))
 //
 
 $lang = array_merge($lang, array(
+	'ACP_GOOGLEANALYTICS'				=> 'Google Analytics',
 	'ACP_GOOGLEANALYTICS_ID'			=> 'Cod de urmarire Google Analytics ',
-	'ACP_GOOGLEANALYTICS_ID_EXPLAIN'	=> 'Scrie codul tau Google Analytics de urmarire, ex: <samp>UA-0000000-00</samp>.',
+	'ACP_GOOGLEANALYTICS_ID_EXPLAIN'	=> 'Scrie codul tau Google Analytics de urmarire, ex: <samp>UA-0000000-00</samp>.<br /><br />Google Analytics poate urmări utilizatorii înregistrați pe mai multe dispozitive și sesiuni, pentru un număr de utilizator mai precise în rapoarte. Pentru a activa această funcționalitate îmbunătățită de utilizare ID-ul de urmărire trebuie să fie configurat în contul dvs. Google Analytics. <a href="https://support.google.com/analytics/answer/3123666">Click pentru mai multe informații</a>.',
 	'ACP_GOOGLEANALYTICS_ID_INVALID'	=> '“%s” nu este un cod valid de urmarire Google Analytics.<br />Trebuie sa fie sub forma “UA-0000000-00”.',
 ));