+++ tests/functional/google_analytics_test.php
@@ -15,6 +15,7 @@ namespace phpbb\googleanalytics\tests\functional;
 */
 class google_analytics_test extends \phpbb_functional_test_case
 {
+	/** @var string */
 	protected $sample_ga_code = 'UA-000000-00';
 
 	/**
@@ -44,11 +45,11 @@ class google_analytics_test extends \phpbb_functional_test_case
 		// Load ACP board settings page
 		$crawler = self::request('GET', 'adm/index.php?i=acp_board&mode=settings&sid=' . $this->sid);
 
-		// Test that GA settings field is found in the correct position (after SYSTEM_TIMEZONE)
+		// Test that GA settings field is found in the correct position (after WARNINGS_EXPIRE)
 		$nodes = $crawler->filter('#acp_board > fieldset > dl > dt > label')->extract(array('_text'));
 		foreach ($nodes as $key => $config_name)
 		{
-			if (strpos($config_name, $this->lang('SYSTEM_TIMEZONE')) !== 0)
+			if (strpos($config_name, $this->lang('WARNINGS_EXPIRE')) !== 0)
 			{
 				continue;
 			}